Components: Split Header Into Two Sub-Components and Extracted One Sub-Component from Main
