+++ src/App.jsx
@@ -1,7 +1,7 @@
 import './App.css'
-import Footer from './components/Footer';
-import Header from './components/Header';
-import Main from './components/Main';
+import Footer from './components/footer/Footer';
+import Header from './components/header/Header';
+import Main from './components/main/Main';
 
 function App() {
   return (
