+++ src/components/header/Header.jsx
@@ -1,0 +1,13 @@
+import HeaderLeftMenu from "./HeaderLeftMenu"
+import HeaderRightMenu from "./HeaderRightMenu"
+
+function Header() {
+    return (
+        <header>
+            <HeaderLeftMenu/>
+            <HeaderRightMenu/>
+        </header>
+    )
+}
+
+export default Header
