+++ src/components/header/HeaderLeftMenu.jsx
@@ -1,0 +1,13 @@
+function HeaderLeftMenu() {
+    return (
+        <nav className="left-menu">
+        <ul>
+          <li>About</li>
+          <li>Store</li>
+        </ul>
+      </nav>
+
+    )
+}
+
+export default HeaderLeftMenu
