+++ src/components/main/Main.jsx
@@ -1,3 +1,5 @@
+import Search from "./Search"
+
 function Main() {
     return (
         <main>
@@ -6,11 +8,7 @@ function Main() {
           src="https://www.google.com/images/branding/googlelogo/2x/googlelogo_color_272x92dp.png"
           alt="Google logo"
         />
-        <input className="search-bar" type="search" />
-        <div className="buttons">
-          <button>Google Search</button>
-          <button>Feeling Lucky</button>
-        </div>
+        <Search />
       </main>
     )
 }
