+++ src/components/main/Search.jsx
@@ -1,0 +1,12 @@
+function Search() {
+    return (
+        <>
+            <input className="search-bar" type="search" />
+            <div className="buttons">
+            <button>Google Search</button>
+            <button>Feeling Lucky</button>
+            </div>
+        </>
+    )
+}
+export default Search
